backend/requirements.txt
python-dotenv
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1